watch/serializers.py
from rest_framework import serializers
from .models import *

class PostsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Posts
        fields = '__all__'

class BusinessSerializer(serializers.ModelSerializer):
    class Meta:
        model = Business
        fields= ('bsn_name','bsn_email', 'user')