requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
Django==4.0.5
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.50
Pillow==9.1.1
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
